Extract objects to files
* periods
* situations
* reforms
